[WEBSITES-537] runtime formProxy utility (NewsletterForm)
