+++ bff.js
@@ -1,9 +1,9 @@
 const axios = require('axios').default;
-const React = require('react');
 const fs = require('fs');
 const path = require('path');
 const sh = require('shelljs');
-const pmUtilities = ['sanatizeContent'];
+
+const pmUtilities = ['sanatizeContent', 'proxyForm'];
 
 const host = 'https://postman-web-property-assets.s3.amazonaws.com';
 
@@ -59,45 +59,57 @@ function cachePm() {
 
 axios.get(`${host}/api/manifest.json`)
   .then((response) => {
-  	bff = { ...response.data, host };
-  	bffUrl = `${bff.host}/${bff.path}`;
-	cacheDirName = `cache-${bff.path.replace(/(api|bff)\//, '').replace('/', '_')}`;
-	cacheDir = path.join('public', cacheDirName);
-	cachePmDir = path.join(cacheDir, 'pm');
-	script = `
-	  var d = 1000, t, z;
-	  load('/${cacheDirName}/sh.js');
-	  loadPms(${JSON.stringify(pmUtilities)});
-	  if (!z) {
-	    clearTimeout(t);
-	    t = setTimeout(function(){
-	      if (window.pm && window.pm.sanatizeContent) {
-	        window.pm.cache = '${cacheDirName}';
-	        clearTimeout(t);
-	      }
-	    }, d);
-	  }
-	  function load(src) {
-	    var e = document.createElement('script');
-	    e.src = src;
-	    document.head.appendChild(e);
-	  }
-	  function loadPm(name) {
-	    load('/${cacheDirName}/pm/' + name + '.js');
-	  }
-	  function loadPms(names) {
-	    var i, max = names.length;
-	    for (i = 0; i < max; i++) {
-	      loadPm(names[i]);
-	    }
-	  }
-	`;	
+    bff = { ...response.data, host };
+    bffUrl = `${bff.host}/${bff.path}`;
+    cacheDirName = `cache-${bff.path.replace(/(api|bff)\//, '').replace('/', '_')}`;
+    cacheDir = path.join('public', cacheDirName);
+    cachePmDir = path.join(cacheDir, 'pm');
+    script = `
+    var d = 1000, t, z;
+    load('/${cacheDirName}/sh.js');
+    loadPms(${JSON.stringify(pmUtilities)});
+    if (!z) {
+      clearTimeout(t);
+      t = setTimeout(function(){
+        if (window.pm && window.pm.sanatizeContent) {
+          window.pm.cache = '${cacheDirName}';
+          clearTimeout(t);
+        }
+      }, d);
+    }
+    window.lazyLoadPmUtility = function(name, cb, optionalDelay) {
+      var delay = optionalDelay || 1000;
+      setTimeout(function(){
+        var hasDependencies = window && window.pm && window.pm.cache;
+        if (hasDependencies) {
+          const e = document.createElement('script');
+          e.src = '/' + window.pm.cache + '/pm/' + name + '.js';
+          e.onload = cb;
+          document.head.appendChild(e);
+        }
+      }, delay);
+    }
+    function load(src) {
+      var e = document.createElement('script');
+      e.src = src;
+      document.head.appendChild(e);
+    }
+    function loadPm(name) {
+      load('/${cacheDirName}/pm/' + name + '.js');
+    }
+    function loadPms(names) {
+      var i, max = names.length;
+      for (i = 0; i < max; i++) {
+        loadPm(names[i]);
+      }
+    }
+  `;
 
     fs.writeFile('bff.json', JSON.stringify({ ...response.data, host, script }), (err) => {
       if (err) {
         throw err;
       } else {
-      	cachePm();
+        cachePm();
       }
     });
   });
