+++ src/utils/proxyForm.jsx
@@ -1,0 +1,34 @@
+import React from 'react';
+
+const proxyForm = ({
+  track, id, className, description, form,
+}) => {
+  if (typeof document === 'object' && track && id) {
+    window.lazyLoadPmUtility('proxyForm', () => {
+      window.pm.proxyForm({
+        host: '//pages.getpostman.com',
+        track,
+        id,
+      });
+    });
+
+    return (
+      <div className={className}>
+        {description}
+        {form}
+        <form id={id} />
+        <style>
+          {`
+          #${id} {
+            display: none !important;
+          }
+        `}
+        </style>
+      </div>
+    );
+  }
+
+  return null;
+};
+
+export default proxyForm;
